Migrate thread functions to rsbs/src/libultra/os/ (WP-009 Batch 4)
Migrate identical/functionally-identical thread-related functions from
games/*/src/libultra/os/ to the unified rsbs library:

- getactivequeue.c: Returns __osActiveQueue (identical)
- getthreadid.c: Returns thread->id with NULL check (functionally identical)
- getthreadpri.c: Returns thread->priority with NULL check (functionally identical)
- destroythread.c: Removes thread from queues (functionally identical)
- dequeuethread.c: Removes thread from specified queue (OoT only)

Note: createthread.c was NOT migrated because OoT and MM have different
implementations (OoT includes global variable definitions and uses
different constants/signatures).

Add rsbs/include/rsbs/libultra_os.h header providing necessary type
declarations and externs for the unified libultra/os functions.
