--- games/mm/src/libultra/os/getthreadid.c
@@ -1,8 +1,0 @@
-#include "ultra64.h"
-
-OSId osGetThreadId(OSThread* t) {
-    if (t == NULL) {
-        t = __osRunningThread;
-    }
-    return t->id;
-}
