--- games/mm/src/libultra/os/getthreadpri.c
@@ -1,8 +1,0 @@
-#include "ultra64.h"
-
-OSPri osGetThreadPri(OSThread* t) {
-    if (t == NULL) {
-        t = __osRunningThread;
-    }
-    return t->priority;
-}
