--- games/oot/src/libultra/os/getthreadid.c
@@ -1,9 +1,0 @@
-#include "global.h"
-
-OSId osGetThreadId(OSThread* thread) {
-    if (thread == NULL) {
-        thread = __osRunningThread;
-    }
-
-    return thread->id;
-}
