--- games/oot/src/libultra/os/getthreadpri.c
@@ -1,9 +1,0 @@
-#include "global.h"
-
-OSPri osGetThreadPri(OSThread* thread) {
-    if (thread == NULL) {
-        thread = __osRunningThread;
-    }
-
-    return thread->priority;
-}
